+++ lib/response.ml
@@ -41,7 +41,11 @@ type t =
  *   HTTP/2 does not define a way to carry the version or reason phrase that
  *   is included in an HTTP/1.1 status line. *)
 let create ?(headers = Headers.empty) status = { status; headers }
-let body_length { headers; _ } = Message.body_length headers
+
+let body_length ~request_method { headers; _ } =
+  match request_method with
+  | `HEAD -> `Fixed 0L
+  | #Httpaf.Method.standard -> Message.body_length headers
 
 let pp_hum fmt { status; headers } =
   let reason =